Adding function for connection string
