+++ .gitignore
@@ -1,2 +1,3 @@
 main
 .DS_Store
+images/tmp
